Update for gtkmm 3 and 4 instead of 2 and 3
Many updates. Some examples:
Links suitable for gtkmm 4 and to some extent gtkmm 3.
The News page is totally rewritten.
GTK+ is changed to GTK.
Many http:// are changed to https://.
The license is called Lesser GPL (was Library GPL).
gtk-list is not used nowadays. Link to discourse.gnome.org instead.